requirements.txt
crewai
langchain
openai
jinja2
streamlit
python-dotenv
requests